PR30 - Componente para visualização de itens
