+++ package.json
@@ -3,14 +3,15 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@emotion/react": "^11.9.3",
-    "@emotion/styled": "^11.9.3",
+    "@emotion/react": "^11.10.5",
+    "@emotion/styled": "^11.10.5",
     "@mantine/core": "^5.7.2",
     "@mantine/hooks": "^5.7.2",
     "@mantine/next": "^5.7.2",
     "@material-ui/core": "^4.12.4",
     "@material-ui/icons": "^4.11.3",
-    "@mui/material": "^5.8.4",
+    "@mui/icons-material": "^5.10.16",
+    "@mui/material": "^5.10.16",
     "@mui/styles": "^5.8.4",
     "@progress/kendo-react-buttons": "^5.9.0",
     "@progress/kendo-react-data-tools": "^5.9.0",
@@ -26,25 +27,30 @@
     "@testing-library/user-event": "^13.5.0",
     "axios": "^0.27.2",
     "bootstrap": "^5.2.2",
+    "crypto-js": "^3.1.9-1",
     "goober": "^2.1.11",
     "install": "^0.13.0",
     "jwt-decode": "^3.1.2",
     "mdb-react-ui-kit": "^4.2.0",
     "moment": "^2.29.4",
+    "prop-types": "^15.8.1",
     "react": "^18.2.0",
     "react-bootstrap": "^2.5.0",
     "react-dom": "^18.2.0",
-    "react-icons": "^4.6.0",
+    "react-icons": "^4.7.1",
     "react-input-mask": "^2.0.4",
     "react-multistep": "^5.3.2",
     "react-paginate": "^8.1.4",
     "react-rounded-image": "^2.0.14",
     "react-router-dom": "^6.3.0",
     "react-scripts": "^5.0.1",
+    "react-table-6": "^6.11.0",
+    "react-table-css-unit": "^0.1.1",
     "reactstrap": "^9.1.4",
     "sass": "^1.55.0",
     "semantic-ui-react": "^2.1.3",
     "styled-components": "^5.3.5",
+    "tailwind": "^4.0.0",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
