--- src/components/Box/DefaultBox.js
@@ -1,16 +1,0 @@
-import styled from 'styled-components';
-
-//mudar nome da box abaixo
-
-export const DefaultBox = styled.div`
-background-color: #FFFFFF;
-border-radius: 15px;
-width: ${props => props.width}px;
-height: ${props => props.height}px;
-align-items: center;
-justify-content: center;
-border: solid 1px;
-border-color: #0E6BA8;
-box-shadow: 0 0 5px #03112C;
-display: ${props => props.display};
-`;
