+++ src/components/Structured/Modal/styles.js
