+++ src/components/Structured/Pagination/Paginate.jsx
@@ -1,0 +1,62 @@
+import React, { useEffect, useState } from 'react';
+import ReactPaginate from 'react-paginate';
+
+const defaultItems = [...Array(1).keys()];
+
+function Items({ currentItems }) {
+  return (
+    <div className="items">
+    {currentItems && currentItems.map((item) => (
+      <div>
+        <h3>Item #{item}</h3>
+      </div>
+    ))}
+      </div>
+  );
+}
+
+export function PaginatedItems({ itemsPerPage, itens }) {
+  const [currentItems, setCurrentItems] = useState(null);
+  const [pageCount, setPageCount] = useState(0);
+  const [itemOffset, setItemOffset] = useState(0);
+  const itensPaginated = itens ?? defaultItems;
+
+  useEffect(() => {
+    const endOffset = itemOffset + itemsPerPage;
+    console.log(`Loading items from ${itemOffset} to ${endOffset}`);
+    setCurrentItems(itensPaginated.slice(itemOffset, endOffset));
+    setPageCount(Math.ceil(itensPaginated.length / itemsPerPage));
+  }, [itemOffset, itemsPerPage]);
+
+  const handlePageClick = (event) => {
+    const newOffset = event.selected * itemsPerPage % itensPaginated.length;
+    console.log(`User requested page number ${event.selected}, which is offset ${newOffset}`);
+    setItemOffset(newOffset);
+  };
+
+  return (
+    <>
+      <Items currentItems={currentItems} />
+      <ReactPaginate
+        nextLabel="next >"
+        onPageChange={handlePageClick}
+        pageRangeDisplayed={3}
+        marginPagesDisplayed={2}
+        pageCount={pageCount}
+        previousLabel="< previous"
+        pageClassName="page-item"
+        pageLinkClassName="page-link"
+        previousClassName="page-item"
+        previousLinkClassName="page-link"
+        nextClassName="page-item"
+        nextLinkClassName="page-link"
+        breakLabel="..."
+        breakClassName="page-item"
+        breakLinkClassName="page-link"
+        containerClassName="pagination"
+        activeClassName="active"
+        renderOnZeroPageCount={null}
+      />
+    </>
+  );
+}
