+++ src/components/Structured/Table/styles.css
@@ -1,0 +1,9 @@
+.react-table-style {
+  height: 470px;
+  width: 1300px;
+  border: #2A2C35 0px solid;
+  box-shadow: 0 0 5px #2A2C35;
+  border-radius: 10px;
+  margin-top: 1%;
+  margin-left: 7%;
+}
