--- src/components/Submenu/Submenu.jsx
@@ -1,36 +1,0 @@
-import React, { useState } from 'react';
-
-import { SidebarLink, SidebarLabel, DropdownLink } from "./submenu"
-
-export function SubMenu({ item }){
-  const [subnav, setSubnav] = useState(false);
-
-  const showSubnav = () => setSubnav(!subnav);
-
-  return (
-    <div>
-      <SidebarLink to={item.path} onClick={item.subNav && showSubnav}>
-        <div>
-          {item.icon}
-          <SidebarLabel>{item.title}</SidebarLabel>
-          {
-            item.subNav && subnav
-            ? item.iconOpened
-            : item.subNav
-            ? item.iconClosed
-            : null
-          }
-        </div>
-      </SidebarLink>
-      {subnav &&
-        item.subNav.map((item, index) => {
-          return (
-            <DropdownLink to={item.path} key={index}>
-              {item.icon}
-              <SidebarLabel>{item.title}</SidebarLabel>
-            </DropdownLink>
-          );
-        })}
-    </div>
-  );
-};
