+++ src/services/Getters/users.js
@@ -1,0 +1,7 @@
+import { createAPIEndpoint, ENDPOINTS } from "../apiMap";
+
+export const handleGetAllUsersInfo = () => {
+
+    return createAPIEndpoint(ENDPOINTS.others.getAllUsers)
+           .getAll();
+}
